Implement preview while drawing and render box drawing characters
The grid buffer is now a vector of GridCell, a structure that not only
holding cell character, but also the preview content.

The way it works is: While drawing, we set the preview character for each
cell, then, depends on the next action (finish or continue drawing), we
commit() or discard() the preview character.